requirements.txt
requires.io
pillow
pybase64
requests
jsons
python-dotenv